자잘한 수정

+++ src/main/java/com/playjo/project/playjo/domain/Notice.java
@@ -19,7 +19,7 @@ public class Notice {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int notice_no;
+    private Long notice_no;
 
     @Lob
     private String content;
